BI-14526: Move common config to common/config package
