+++ src/main/java/uk/gov/companieshouse/officerssearch/subdelta/common/config/AppConfig.java
@@ -1,0 +1,34 @@
+package uk.gov.companieshouse.officerssearch.subdelta.common.config;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import java.util.function.Supplier;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import uk.gov.companieshouse.api.InternalApiClient;
+import uk.gov.companieshouse.api.http.ApiKeyHttpClient;
+
+@Configuration
+public class AppConfig {
+
+    @Bean
+    public ObjectMapper objectMapper() {
+        return new ObjectMapper()
+                .setSerializationInclusion(JsonInclude.Include.NON_NULL)
+                .registerModule(new JavaTimeModule());
+    }
+
+    @Bean
+    Supplier<InternalApiClient> internalApiClientSupplier(
+            @Value("${api.api-key}") String apiKey,
+            @Value("${api.api-url}") String apiUrl) {
+        return () -> {
+            InternalApiClient internalApiClient = new InternalApiClient(new ApiKeyHttpClient(
+                    apiKey));
+            internalApiClient.setBasePath(apiUrl);
+            return internalApiClient;
+        };
+    }
+}
